third_party/WebKit/LayoutTests/external/wpt/fetch/cors-rfc1918/idlharness.tentative.any.worker-expected.txt

This is a testharness.js-based test.
PASS Test CORS RFC1918 interfaces
PASS Partial interface Document: original interface defined
PASS Partial interface WorkerGlobalScope: original interface defined
FAIL WorkerGlobalScope interface: attribute addressSpace assert_true: The prototype object must have a property "addressSpace" expected true got false
PASS Unscopable handled correctly for addressSpace property on WorkerGlobalScope
FAIL WorkerGlobalScope interface: self must inherit property "addressSpace" with the proper type assert_inherits: property "addressSpace" not found in prototype chain
PASS Node interface: existence and properties of interface object
PASS Document interface: existence and properties of interface object
FAIL Document interface: document must not have property "addressSpace" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: document is not defined"
Harness: the test ran to completion.
